Ändringar i MovieVC, UpdateFilmC, Film, FilmM

+++ src/main/java/org/dreamteam/wigellsmoviesstore/Entitys/Film.java
@@ -116,7 +116,7 @@ public class Film {
         this.description = description;
     }
 
-    public int getReleaseYear() {
+    public short getReleaseYear() {
         return releaseYear;
     }
